% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkinfer-package.R
\docType{package}
\name{linkinfer-package}
\alias{linkinfer}
\alias{linkinfer-package}
\title{linkinfer: Methods for Inference with Linked Data}
\description{
Tools for diagnosing and correcting bias arising from record linkage errors. Provides methods for comparing linked samples to reference populations, adjusting for missed matches using inverse probability weighting, post-stratification, and raking, and validating adjustments using validation variables. Companion package to Breen and Joo (2026) "Population Research With Linked Data: Methods for Inference".
}
\author{
\strong{Maintainer}: Casey F. Breen \email{casey.breen@austin.utexas.edu}

Authors:
\itemize{
  \item Won-tak Joo \email{wjoo@ufl.edu}
}

}
\keyword{internal}
